Moved FilePickerExampleActivity to a example module
