+++ app/src/main/AndroidManifest.xml
@@ -6,9 +6,6 @@
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
 
     <application>
-        <activity
-            android:name=".FilePickerExampleActivity"
-            android:screenOrientation="portrait" />
 
         <activity
             android:name=".FilePickerActivity"
